feat: run listeners to details
