+++ ui/src/main/java/com/cairosquad/ui/home/HomeScreen.kt
@@ -9,10 +9,9 @@ import com.cairosquad.ui.navigation.LocalNavController
 import com.cairosquad.ui.navigation.MovieRoute
 import com.cairosquad.ui.navigation.SeriesRoute
 import com.cairosquad.ui.utils.ObserveAsEffect
-import com.cairosquad.viewmodel.home.effect.DiscoverEffect
 import com.cairosquad.viewmodel.home.DiscoverViewModel
-import com.cairosquad.viewmodel.home.effect.HomeEffect
 import com.cairosquad.viewmodel.home.HomeViewModel
+import com.cairosquad.viewmodel.home.effect.HomeEffect
 import com.cairosquad.viewmodel.home.model.DiscoverType
 import org.koin.androidx.compose.koinViewModel
 
@@ -24,12 +23,6 @@ fun HomeScreen(
 
     val navController = LocalNavController.current
     ObserveAsEffect(viewModel.effect) { effect -> effectHandler(effect, navController) }
-    ObserveAsEffect(discoverViewModel.effect) { effect ->
-        effectDiscoverHandler(
-            effect, navController
-        )
-    }
-
     val screenState by viewModel.screenState.collectAsState()
 
     HomeScreenContent(
@@ -50,46 +43,30 @@ private fun effectHandler(
         is HomeEffect.NavigateSeries -> {
             navController.navigate(SeriesRoute(effect.seriesId))
         }
+        is HomeEffect.NavigateToDiscover -> {
+            navController.navigate("discover/${effect.type.name}/${effect.mediaType.name}")
+        }
+
+        is HomeEffect.NavigateToSeeAllFreeToWatch -> {
+            navController.navigate("discover/${DiscoverType.FREE_TO_WATCH.name}/${effect.mediaType.name}")
+        }
+
+        is HomeEffect.NavigateToSeeAllTopRating -> {
+            navController.navigate("discover/${DiscoverType.TOP_RATING.name}/${effect.mediaType.name}")
+        }
+
+        is HomeEffect.NavigateToSeeAllUpcoming -> {
+            navController.navigate("discover/${DiscoverType.UPCOMING.name}/${effect.mediaType.name}")
+        }
+
+        is HomeEffect.NavigateToSeeAllTrending -> {
+            navController.navigate("discover/${DiscoverType.TRENDING.name}/${effect.mediaType.name}")
+        }
 
         HomeEffect.NavigateToProfile -> { /* TODO: Navigate to profile */
         }
 
         HomeEffect.NavigateToSeeAllAiringToday -> {}
         HomeEffect.NavigateToSeeAllOnTv -> {}
-    }
-}
-
-private fun effectDiscoverHandler(
-    effect: DiscoverEffect, navController: NavController
-) {
-
-    when (effect) {
-        is DiscoverEffect.NavigateMovie -> {
-            navController.navigate(MovieRoute(effect.movieId))
-        }
-
-        is DiscoverEffect.NavigateSeries -> {
-            navController.navigate(SeriesRoute(effect.seriesId))
-        }
-
-        is DiscoverEffect.NavigateToDiscover -> {
-            navController.navigate("discover/${effect.type.name}/${effect.mediaType.name}")
-        }
-
-        is DiscoverEffect.NavigateToSeeAllFreeToWatch -> {
-            navController.navigate("discover/${DiscoverType.FREE_TO_WATCH.name}/${effect.mediaType.name}")
-        }
-
-        is DiscoverEffect.NavigateToSeeAllTopRating -> {
-            navController.navigate("discover/${DiscoverType.TOP_RATING.name}/${effect.mediaType.name}")
-        }
-
-        is DiscoverEffect.NavigateToSeeAllUpcoming -> {
-            navController.navigate("discover/${DiscoverType.UPCOMING.name}/${effect.mediaType.name}")
-        }
-
-        is DiscoverEffect.NavigateToSeeAllTrending -> {
-            navController.navigate("discover/${DiscoverType.TRENDING.name}/${effect.mediaType.name}")
-        }
     }
 }
