+++ ui/src/main/java/com/cairosquad/ui/home/content/HomeScreenContentSeriesTab.kt
@@ -15,16 +15,14 @@ import com.cairosquad.ui.movio_component.MediaHorizontalPagerItem
 import com.cairosquad.ui.movio_component.MediaSection
 import com.cairosquad.ui.movio_component.MediaSectionItem
 import com.cairosquad.ui.movio_component.MediaSectionLayoutType
-import com.cairosquad.viewmodel.home.listner.DiscoverInteractionsListener
 import com.cairosquad.viewmodel.home.listner.HomeInteractionsListener
-import com.cairosquad.viewmodel.home.state.HomeScreenState
 import com.cairosquad.viewmodel.home.model.MediaType
+import com.cairosquad.viewmodel.home.state.HomeScreenState
 
 @Composable
 fun HomeScreenContentSeriesTab(
     screenState: HomeScreenState,
     listener: HomeInteractionsListener,
-    discoverInteractionsListener: DiscoverInteractionsListener,
     scrollState: ScrollState
 ) {
     Column(
@@ -50,7 +48,7 @@ fun HomeScreenContentSeriesTab(
             },
             sectionTitle = stringResource(R.string.top_rating),
             mediaSectionLayoutType = MediaSectionLayoutType.LazyRow,
-            seeAllAction = { discoverInteractionsListener.onClickSeeAllTopRated(MediaType.Movies) }
+            seeAllAction = { listener.onClickSeeAllTopRated(MediaType.Movies) }
         )
 
         MediaSection(
@@ -88,7 +86,7 @@ fun HomeScreenContentSeriesTab(
             },
             sectionTitle = stringResource(R.string.more_recommended),
             mediaSectionLayoutType = MediaSectionLayoutType.LazyVerticalGrid(158),
-            seeAllAction = { discoverInteractionsListener.onClickSeeAllMoreRecommended(MediaType.Movies) }
+            seeAllAction = { listener.onClickSeeAllMoreRecommended(MediaType.Movies) }
         )
     }
 }
